Move public functions to `celoutil.go`
